Flatten the structure of src by moving clipping.py one level up
