+++ src/clipping.py
